// src/common/enums/TextureKeys.ts
export enum TextureKeys {
  BUTTON_START = 'button_start',
  BUTTON_START_PRESED = 'button_start_presed',
  BUTTON_GET = 'button_get',
  BUTTON_GET_PRESED = 'button_get_presed',
  BUTTON_STOP = 'button_stop',
  BUTTON_STOP_PRESED = 'button_stop_presed',
  BET_IMAGE = 'bet_image',
  BET_MINUS = 'bet_minus',
  BET_PLUS = 'bet_plus',
}